Support Neo4j.Driver v4.0
