+++ src/IntegrationTests/Models/TimeStamp.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 
 namespace IntegrationTests.Models
 {
