+++ src/Neo4jMapper/RecordExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 
 namespace Neo4jMapper
 {
