+++ src/Neo4jMapper/ResultCursorExtensions.cs
@@ -1,267 +1,267 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 
 namespace Neo4jMapper
 {
-    public static class StatementResultCursorExtensions
+    public static class ResultCursorExtensions
     {
         public static async Task<List<TReturn>> MapAsync<TReturn>(
-            this IStatementResultCursor result)
+            this IResultCursor resultCursor)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map<TReturn>());
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TValue16, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TValue16, TReturn> mapFunc)
         {
-            return await result.MapAsync(
+            return await resultCursor.MapAsync(
                 record => record.Map(mapFunc));
         }
 
         public static async Task<List<TReturn>> MapAsync<TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<IRecord, TReturn> mapFunc)
         {
             var list = new List<TReturn>();
-            while (await result.FetchAsync().ConfigureAwait(false))
+            while (await resultCursor.FetchAsync().ConfigureAwait(false))
             {
-                list.Add(mapFunc(result.Current));
+                list.Add(mapFunc(resultCursor.Current));
             }
             return list;
         }
 
         public static async Task<TReturn> MapSingleAsync<TReturn>(
-            this IStatementResultCursor result)
+            this IResultCursor resultCursor)
         {
-            return (await result.SingleAsync()).Map<TReturn>();
+            return (await resultCursor.SingleAsync()).Map<TReturn>();
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
         public static async Task<TReturn> MapSingleAsync<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TValue16, TReturn>(
-            this IStatementResultCursor result,
+            this IResultCursor resultCursor,
             Func<TValue1, TValue2, TValue3, TValue4, TValue5, TValue6, TValue7, TValue8, TValue9, TValue10, TValue11, TValue12, TValue13, TValue14, TValue15, TValue16, TReturn> mapFunc)
         {
-            return (await result.SingleAsync()).Map(mapFunc);
+            return (await resultCursor.SingleAsync()).Map(mapFunc);
         }
 
-        public static async IAsyncEnumerable<IRecord> AsyncResults(this IStatementResultCursor result)
+        public static async IAsyncEnumerable<IRecord> AsyncResults(this IResultCursor resultCursor)
         {
-            while (await result.FetchAsync())
+            while (await resultCursor.FetchAsync())
             {
-                yield return result.Current;
+                yield return resultCursor.Current;
             }
         }
     }
