+++ src/Neo4jMapper/ValueMapper.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 using ServiceStack;
 
 namespace Neo4jMapper
