+++ src/UnitTests/DictionaryExtensionsTests.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 using Neo4jMapper;
 using NUnit.Framework;
