+++ src/UnitTests/Models/TemporalTypeModels.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 
 namespace UnitTests.Models
 {
