+++ src/UnitTests/RecordExtensionsTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Neo4j.Driver.V1;
+using Neo4j.Driver;
 using Neo4jMapper;
 using NSubstitute;
 using NUnit.Framework;
@@ -613,7 +613,7 @@ namespace UnitTests
             var record = Substitute.For<IRecord>();
             record[0].Returns(1);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -631,7 +631,7 @@ namespace UnitTests
             record[0].Returns(1);
             record[1].Returns(2);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -655,7 +655,7 @@ namespace UnitTests
             record[1].Returns(2);
             record[2].Returns(3);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -682,7 +682,7 @@ namespace UnitTests
             record[2].Returns(3);
             record[3].Returns(4);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -712,7 +712,7 @@ namespace UnitTests
             record[3].Returns(4);
             record[4].Returns(5);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -745,7 +745,7 @@ namespace UnitTests
             record[4].Returns(5);
             record[5].Returns(6);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -782,7 +782,7 @@ namespace UnitTests
             record[5].Returns(6);
             record[6].Returns(7);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -822,7 +822,7 @@ namespace UnitTests
             record[6].Returns(7);
             record[7].Returns(8);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -865,7 +865,7 @@ namespace UnitTests
             record[7].Returns(8);
             record[8].Returns(9);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -911,7 +911,7 @@ namespace UnitTests
             record[8].Returns(9);
             record[9].Returns(10);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -960,7 +960,7 @@ namespace UnitTests
             record[9].Returns(10);
             record[10].Returns(11);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -1012,7 +1012,7 @@ namespace UnitTests
             record[10].Returns(11);
             record[11].Returns(12);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -1068,7 +1068,7 @@ namespace UnitTests
             record[11].Returns(12);
             record[12].Returns(13);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -1127,7 +1127,7 @@ namespace UnitTests
             record[12].Returns(13);
             record[13].Returns(14);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -1189,7 +1189,7 @@ namespace UnitTests
             record[13].Returns(14);
             record[14].Returns(15);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
 
@@ -1254,7 +1254,7 @@ namespace UnitTests
             record[14].Returns(15);
             record[15].Returns(16);
 
-            var records = Substitute.For<IStatementResult>();
+            var records = Substitute.For<IResult>();
             records.GetEnumerator().Returns(
                 new List<IRecord> {record, record}.GetEnumerator());
